pluralize mutations
