+++ packages/authx/src/graphql/mutation/index.ts
@@ -1,43 +1,43 @@
-import { createClient } from "./createClient";
-import { updateClient } from "./updateClient";
+import { createClients } from "./createClients";
+import { updateClients } from "./updateClients";
 import { GraphQLCreateClientInput } from "./GraphQLCreateClientInput";
 import { GraphQLUpdateClientInput } from "./GraphQLUpdateClientInput";
 
-import { createGrant } from "./createGrant";
-import { updateGrant } from "./updateGrant";
+import { createGrants } from "./createGrants";
+import { updateGrants } from "./updateGrants";
 import { GraphQLCreateGrantInput } from "./GraphQLCreateGrantInput";
 import { GraphQLUpdateGrantInput } from "./GraphQLUpdateGrantInput";
 
-import { createRole } from "./createRole";
-import { updateRole } from "./updateRole";
+import { createRoles } from "./createRoles";
+import { updateRoles } from "./updateRoles";
 import { GraphQLCreateRoleInput } from "./GraphQLCreateRoleInput";
 import { GraphQLUpdateRoleInput } from "./GraphQLUpdateRoleInput";
 
-import { createAuthorization } from "./createAuthorization";
-import { updateAuthorization } from "./updateAuthorization";
+import { createAuthorizations } from "./createAuthorizations";
+import { updateAuthorizations } from "./updateAuthorizations";
 import { GraphQLCreateAuthorizationInput } from "./GraphQLCreateAuthorizationInput";
 import { GraphQLUpdateAuthorizationInput } from "./GraphQLUpdateAuthorizationInput";
 
-import { createUser } from "./createUser";
-import { updateUser } from "./updateUser";
+import { createUsers } from "./createUsers";
+import { updateUsers } from "./updateUsers";
 import { GraphQLCreateUserInput } from "./GraphQLCreateUserInput";
 import { GraphQLUpdateUserInput } from "./GraphQLUpdateUserInput";
 
 export const mutationFields = {
-  createClient,
-  updateClient,
+  createClients,
+  updateClients,
 
-  createGrant,
-  updateGrant,
+  createGrants,
+  updateGrants,
 
-  createRole,
-  updateRole,
+  createRoles,
+  updateRoles,
 
-  createAuthorization,
-  updateAuthorization,
+  createAuthorizations,
+  updateAuthorizations,
 
-  createUser,
-  updateUser
+  createUsers,
+  updateUsers
 };
 
 export const mutationTypes = [
